(feat) Implement a registration and eventing abstraction over rust-event.
The new system is based upon three traits:
  - IoHandle
  - Configured
  - Evented

As well as (later) two reading/writing traits, which provide failing
(in the case of an operation which might block) reads and writes.
  - Read
  - Write

IoHandle and Configured are used for registration of handles on the
event loop through the lower-level system. IoHandle provides the
IoDesc that uniquely identifies an IO resource to be registered.
Configured provides the interest and pollopt options to be used
when listening for events on the IoDesc provided by IoHandle.

Evented separately provides the handlers to be called when reads and
writes can be performed without blocking. It's key to the design that
Evented is implementable completely separately from IoHandle and
Configured, so that it can be used for evented APIs that are not
OS IO objects for which an IoDesc can be created.

Read and Write provide the low-level utilities to actually read and
write within the callbacks registered on Evented. The operations
on Read and Write may fail with a WouldBlock AioError, but are
guaranteed to not fail if they are called from an Evented callback.

Registration then provides a convenient wrapper for creating Evented,
and sometimes IoHandle/Configured objects. Registration's contain
configuration, but *can* be used without them. It's likely that aio
will grow a separate pair of IoRegistration and Registration structs,
and only IoRegistration will contain IO-related configuration.